current-tweaks (#15)
* Add Orientation.json template for sexual orientation roles

* Disable report and statistics buttons

Commented out the EmployeeReportButton and EmployeeStatisticsButton in EmployeeManagementMenuView. This change is likely temporary, possibly due to pending updates or issues with these features.

* Add character removal functionality

Implemented the ability to remove staff characters from the system. Added user confirmation and validation steps before deletion to ensure operational integrity. Updated relevant classes to support this feature.

* Add Data Center template for Reaction Roles

* Update set_num_winners to use a dropdown menu

Replaced the numerical input modal with a dropdown menu for selecting the number of winners. This change enhances UX by providing a clearer and more structured way for users to set the number of winners.